+++ src/scout_apm/falcon.py
@@ -1,6 +1,7 @@
 # coding=utf-8
 from __future__ import absolute_import, division, print_function, unicode_literals
 
+import inspect
 import logging
 
 import falcon
@@ -26,15 +27,37 @@ class ScoutMiddleware(object):
     """
 
     def __init__(self, config):
-        install(config=config)
         self.api = None
+        self._attempted_to_discover_api = False
+        installed = install(config=config)
+        self._do_nothing = not installed
 
     def set_api(self, api):
         if not isinstance(api, falcon.API):
             raise ValueError("api should be an instance of falcon.API")
         self.api = api
 
+    def _discover_api(self):
+        """
+        Discover the Falcon API this middleware is attached to via stack
+        inspection. If it fails, record the fact, so we don't attempt it again.
+        """
+        if self._attempted_to_discover_api:
+            return
+        self._attempted_to_discover_api = True
+        try:
+            frame = inspect.currentframe()
+            process_request_frame = frame.f_back
+            API_call_frame = process_request_frame.f_back
+            self.api = API_call_frame.f_locals["self"]
+        except Exception:  # pragma: no cover
+            pass
+
     def process_request(self, req, resp):
+        if self._do_nothing:
+            return
+        if self.api is None:
+            self._discover_api()
         tracked_request = TrackedRequest.instance()
         tracked_request.is_real_request = True
         req.context.scout_tracked_request = tracked_request
@@ -69,6 +92,9 @@ class ScoutMiddleware(object):
             track_amazon_request_queue_time(amazon_queue_time, tracked_request)
 
     def process_resource(self, req, resp, resource, params):
+        if self._do_nothing:
+            return
+
         tracked_request = getattr(req.context, "scout_tracked_request", None)
         if tracked_request is None:
             # Somehow we didn't start a request - this might occur in
@@ -80,8 +106,9 @@ class ScoutMiddleware(object):
         if self.api is None:
             logger.warning(
                 (
-                    "{}.set_api() should be called before requests begin for "
-                    "more detail"
+                    "Automatic API object discovery failed. Call {}.set_api()"
+                    " before requests begin to enable more detail to be"
+                    " captured."
                 ).format(self.__class__.__name__)
             )
             operation = "Controller/{}.{}.{}".format(
@@ -92,8 +119,12 @@ class ScoutMiddleware(object):
             # current resource but unfortunately not the method being called, hence
             # we have to go through routing again.
             responder, _params, _resource, _uri_template = self.api._get_responder(req)
+            try:
+                last_part = responder.__name__
+            except AttributeError:
+                last_part = req.method
             operation = "Controller/{}.{}.{}".format(
-                resource.__module__, resource.__class__.__name__, responder.__name__
+                resource.__module__, resource.__class__.__name__, last_part
             )
 
         span = tracked_request.start_span(