+++ tests/integration/test_falcon.py
@@ -19,13 +19,21 @@ from tests.integration.util import (
 )
 
 
+class FalconAPI(falcon.API):
+    """
+    Custom subclass that dosen't use __slots__, allowing us to attach
+    'scout_middleware' as an attribute for access in tests.
+    """
+
+    pass
+
+
 @contextmanager
 @kwargs_only
-def app_with_scout(config=None, middleware=None, set_api=True):
+def app_with_scout(config=None, middleware=None):
     """
     Context manager that yields a fresh Falcon app with Scout configured.
     """
-    # Enable Scout by default in tests.
     if config is None:
         config = {}
 
@@ -39,9 +47,7 @@ def app_with_scout(config=None, middleware=None, set_api=True):
     scout_middleware = ScoutMiddleware(config=config)
     middleware[scout_index] = scout_middleware
 
-    app = falcon.API(middleware=middleware)
-    if set_api:
-        scout_middleware.set_api(app)
+    app = FalconAPI(middleware=middleware)
 
     class HomeResource(object):
         def on_get(self, req, resp):
@@ -81,6 +87,17 @@ def app_with_scout(config=None, middleware=None, set_api=True):
 
     app.add_route("/bad-status", BadStatusResource())
 
+    class ResponderClass(object):
+        def __call__(self, req, resp):
+            resp.body = "Responder class"
+
+    class ResponderClassResource(object):
+        on_get = ResponderClass()
+
+    app.add_route("/responder-class", ResponderClassResource())
+
+    app.scout_middleware = scout_middleware
+
     try:
         yield app
     finally:
@@ -114,7 +131,9 @@ def test_home(tracked_requests):
 
 
 def test_home_without_set_api(caplog, tracked_requests):
-    with app_with_scout(set_api=False) as app:
+    with app_with_scout() as app:
+        # Fake that discovery has failed
+        app.scout_middleware._attempted_to_discover_api = True
         response = TestApp(app).get("/")
 
     assert response.status_int == 200
@@ -132,8 +151,9 @@ def test_home_without_set_api(caplog, tracked_requests):
             "scout_apm.falcon",
             logging.WARNING,
             (
-                "ScoutMiddleware.set_api() should be called before requests "
-                "begin for more detail"
+                "Automatic API object discovery failed. Call"
+                " ScoutMiddleware.set_api() before requests begin to enable"
+                " more detail to be captured."
             ),
         )
     ]
@@ -394,3 +414,29 @@ def test_bad_status(tracked_requests):
         span.operation
         == "Controller/tests.integration.test_falcon.BadStatusResource.on_get"
     )
+
+
+def test_responder_class(tracked_requests):
+    with app_with_scout() as app:
+        response = TestApp(app).get("/responder-class")
+
+    assert response.status_int == 200
+    assert response.text == "Responder class"
+    assert len(tracked_requests) == 1
+    tracked_request = tracked_requests[0]
+    assert tracked_request.tags["path"] == "/responder-class"
+    assert tracked_request.active_spans == []
+    assert len(tracked_request.complete_spans) == 1
+    span = tracked_request.complete_spans[0]
+    assert (
+        span.operation
+        == "Controller/tests.integration.test_falcon.ResponderClassResource.GET"
+    )
+
+
+def test_no_monitor(tracked_requests):
+    with app_with_scout(config={"monitor": False}) as app:
+        response = TestApp(app).get("/")
+
+    assert response.status_int == 200
+    assert tracked_requests == []